+++ tests/maddy/test_maddy_orderedlistparser.cpp
@@ -15,10 +15,10 @@ class MADDY_ORDEREDLISTPARSER : public ::testing::Test
 protected:
   std::shared_ptr<maddy::OrderedListParser> olParser;
 
-  void
-  SetUp() override
+  void SetUp() override
   {
-    std::function<std::shared_ptr<maddy::BlockParser>(const std::string& line)> getBlockParserForLineCallback = [](const std::string& line)
+    std::function<std::shared_ptr<maddy::BlockParser>(const std::string& line)>
+      getBlockParserForLineCallback = [](const std::string& line)
     {
       if (maddy::OrderedListParser::IsStartingLine(line))
       {
@@ -32,15 +32,16 @@ protected:
     };
 
     this->olParser = std::make_shared<maddy::OrderedListParser>(
-      nullptr,
-      getBlockParserForLineCallback
+      nullptr, getBlockParserForLineCallback
     );
   }
 };
 
 // -----------------------------------------------------------------------------
 
-TEST_F(MADDY_ORDEREDLISTPARSER, IsStartingLineReturnsTrueWhenFacedWithBeginningOfList)
+TEST_F(
+  MADDY_ORDEREDLISTPARSER, IsStartingLineReturnsTrueWhenFacedWithBeginningOfList
+)
 {
   ASSERT_TRUE(maddy::OrderedListParser::IsStartingLine("1. a"));
 }
@@ -52,11 +53,7 @@ TEST_F(MADDY_ORDEREDLISTPARSER, IsFinishedAlwaysReturnsFalseInTheBeginning)
 
 TEST_F(MADDY_ORDEREDLISTPARSER, ItReplacesMarkdownWithAnHtmlOrderedList)
 {
-  std::vector<std::string> markdown = {
-    "1. a"
-    , "* b"
-    , ""
-  };
+  std::vector<std::string> markdown = {"1. a", "* b", ""};
   std::string expected = "<ol><li>a</li><li>b</li></ol>";
 
   for (std::string md : markdown)
@@ -75,14 +72,11 @@ TEST_F(MADDY_ORDEREDLISTPARSER, ItReplacesMarkdownWithAnHtmlOrderedList)
 TEST_F(MADDY_ORDEREDLISTPARSER, ItReplacesMarkdownWithAnHierachicalHtmlList)
 {
   std::vector<std::string> markdown = {
-    "1. a"
-    , "  1. d"
-    , "  * e"
-    , "* b"
-    , "  1. c"
-    , ""
+    "1. a", "  1. d", "  * e", "* b", "  1. c", ""
   };
-  std::string expected = "<ol><li>a<ol><li>d</li><li>e</li></ol></li><li>b<ol><li>c</li></ol></li></ol>";
+  std::string expected =
+    "<ol><li>a<ol><li>d</li><li>e</li></ol></li><li>b<ol><li>c</li></ol></li></"
+    "ol>";
 
   for (std::string md : markdown)
   {
@@ -97,14 +91,11 @@ TEST_F(MADDY_ORDEREDLISTPARSER, ItReplacesMarkdownWithAnHierachicalHtmlList)
   ASSERT_EQ(expected, outputString);
 }
 
-TEST_F(MADDY_ORDEREDLISTPARSER, ItReplacesNumberedMarkdownListWithAnHtmlOrderedList)
+TEST_F(
+  MADDY_ORDEREDLISTPARSER, ItReplacesNumberedMarkdownListWithAnHtmlOrderedList
+)
 {
-  std::vector<std::string> markdown = {
-    "1. a"
-    , "94. b"
-    , "103. c"
-    , ""
-  };
+  std::vector<std::string> markdown = {"1. a", "94. b", "103. c", ""};
   std::string expected = "<ol><li>a</li><li>b</li><li>c</li></ol>";
 
   for (std::string md : markdown)