+++ tests/main.cpp
@@ -7,7 +7,8 @@
 
 // -----------------------------------------------------------------------------
 
-int main (int argc, char** argv) {
+int main(int argc, char** argv)
+{
   ::testing::GTEST_FLAG(throw_on_failure) = true;
   ::testing::InitGoogleMock(&argc, argv);
   return RUN_ALL_TESTS();